Move `id-factory` to `board`
